+++ lib/twilio-ruby/rest/video/v1/room/room_participant/room_participant_subscribe_rule.rb
@@ -154,7 +154,7 @@ module Twilio
               end
 
               ##
-              # @return [Hash] A collection of Subscribe Rules to include or exclude matching Tracks.
+              # @return [String] A collection of Subscribe Rules to include or exclude matching Tracks.
               def rules
                 @properties['rules']
               end