+++ spec/integration/verify/v2/service/verification_check_spec.rb
@@ -34,7 +34,7 @@ describe 'VerificationCheck' do
           "sid": "VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "service_sid": "VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "to": "+14159373912",
+          "to": "+15017122661",
           "channel": "sms",
           "status": "approved",
           "valid": true,