+++ spec/integration/video/v1/room/room_participant/room_participant_subscribe_rule_spec.rb
@@ -35,7 +35,16 @@ describe 'SubscribeRules' do
           "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_created": null,
           "date_updated": null,
-          "rules": []
+          "rules": [
+              {
+                  "type": "include",
+                  "all": true,
+                  "publisher": null,
+                  "track": null,
+                  "kind": null,
+                  "priority": null
+              }
+          ]
       }
       ]
     ))
@@ -76,7 +85,11 @@ describe 'SubscribeRules' do
           "rules": [
               {
                   "type": "exclude",
-                  "all": true
+                  "all": true,
+                  "publisher": null,
+                  "track": null,
+                  "kind": null,
+                  "priority": null
               }
           ]
       }