+++ CHANGES.md
@@ -1,6 +1,16 @@
 twilio-ruby changelog
 =====================
 
+[2019-07-10] Version 5.25.1
+----------------------------
+**Api**
+- Make `friendly_name` optional for applications create
+- Add new property `as_of` date to Usage Record API calls
+
+**Wireless**
+- Added Usage Records resource.
+
+
 [2019-06-26] Version 5.25.0
 ----------------------------
 **Autopilot**